+++ Budget.Server/Api/Transactions/Models/Responses/TransactionListResponse.cs
@@ -9,7 +9,7 @@ namespace Budget.Server.Api.Transactions.Models.Responses
     {
         public required int Id { get; set; }
         public required TransactionType Type { get; set; }
-        public required double Amount { get; set; }
+        public required decimal Amount { get; set; }
         public required string Reason { get; set; }
         public required DateOnly? Date { get; set; }
     }
